backend_orbyte/usuarios/serializers.py
from rest_framework import serializers
from django.contrib.auth import authenticate


class LoginSerializer(serializers.Serializer):
    email    = serializers.EmailField()
    password = serializers.CharField(write_only=True)

    def validate(self, data):
        email    = data.get('email')
        password = data.get('password')

        user = authenticate(username=email, password=password)

        if not user:
            raise serializers.ValidationError('Credenciales incorrectas.')

        if not user.activo:
            raise serializers.ValidationError('Este usuario está inactivo.')

        data['user'] = user
        return data